Clement candidate decider layout change (#900)
As part of an refresh of our internal tools, we're redesigning the IDOL
Candidate Decider based on my new [figma mocks
here](https://www.figma.com/design/qHNMOp6rrIFMv1wBXsVtYZ/SP25-IDOL-Candidate-Decider-Tool?node-id=122-1577&t=LZFXidMfUZKVlQcB-11).

This first PR is focused on moving all the components around to fit the
new layout of the page.

- First, I removed a few things:
- Removed the Ratings statistics (we don't want to show this anymore for
bias reasons)
- Removed Global progress (we already have the "My progress" progress
bar)
- Removed the "Show other people's votes" (we don't want this because of
bias)
- Removed Candidate ID dropdown (we already have navigation with the
main dropdown)
- Changed the layout of several components:
  - Clearly defined a left and right column 
- Moved "Previous" and "Next" buttons to the sides of the navigation
dropdown
  - Moved the "My progress bar" to the top of the left column
  - Made the Comments section taller
- Grouped the Final selection picker with the Save button at the bottom
of the left column
- Moved the "See applicants name" toggle onto the responses column on
the right
- Moved the Candidate info (name, answers to questions) to the right
column

I also added the first CSS color variables in `App.css` as a preliminary
step to prepare for the rest of the overhaul. They currently aren't used
anywhere but it's to prepare for later!


### A few notes:
- Several of the class names used for styling are temporary. I plan on
making them more semantic.
- Many of the styles (eg: `border: 1px solid #666;` are also temporary
for now and are used for clarity and debugging purposes.
- There remains several accessibility and code cleanliness errors, which
I will get to eventually.


|Before|After|
|-|-|
|<img width="1542" alt="Screenshot 2025-03-27 at 2 45 35 PM"
src="https://github.com/user-attachments/assets/ad14e3ce-75be-41df-8297-1035fcf26804"
/>|<img width="1542" alt="Screenshot 2025-03-27 at 2 45 18 PM"
src="https://github.com/user-attachments/assets/7e58f3df-f29d-4b78-b07b-089fe0ceb0dd"
/>|
