+++ frontend/src/components/Candidate-Decider/ApplicantCredentials.module.css
@@ -3,9 +3,6 @@
   flex-direction: column;
   align-items: left;
   justify-content: left;
-  padding: 20px !important;
-  margin-top: 20px;
-  margin-bottom: 20px;
 }
 
 .iconsContainer {
