+++ frontend/src/components/Candidate-Decider/CandidateDecider.module.css
@@ -3,21 +3,22 @@
 }
 
 .candidateDeciderContainer {
-  padding: 2%;
   display: flex;
+  height: 100%;
+  gap: 32px;
+  padding: 32px;
 }
 
-.applicationContainer {
-  width: 150em;
-  border-right: solid;
-  border-color: lightgrey;
-  padding: 1%;
-  overflow-wrap: break-word;
-  overflow: hidden;
+.leftColumn {
+  width: 500px;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  height: 100%;
+  border: 1px solid #666;
 }
 
 .progressContainer {
-  padding: 1%;
   display: flex;
   width: 100%;
 }
@@ -29,6 +30,8 @@
 }
 
 .searchBar {
+  display: flex;
+  gap: 8px;
   margin-bottom: 1%;
 }
 
@@ -49,4 +52,56 @@
 
 .previousNextButtonContainer {
   margin-right: 0.5% !important;
+}
+
+.responsesContainer {
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+  flex: 1;
+  border: 1px solid #666;
+  padding: 24px;
+}
+
+.top {
+  display: flex;
+  flex-direction: column;
+  flex: 1;
+}
+
+.bottom {
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+}
+
+.saveButtonWrapper {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  padding: 24px;
+  border-top: 1px solid #666;
+}
+
+.saveButtonWrapper button {
+  flex: 1;
+}
+
+.ratingSelectorWrapper {
+  display: flex;
+  flex-direction: column;
+  padding: 24px;
+  border-top: 1px solid #666;
+}
+
+.commentEditorWrapper {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  padding: 24px;
+  border-top: 1px solid #666;
+}
+
+.navigation {
+  padding: 24px;
 }
