+++ frontend/src/components/Candidate-Decider/CandidateDecider.tsx
@@ -1,9 +1,8 @@
-import { useEffect, useState } from 'react';
-import { Button, Dropdown, Checkbox, Modal } from 'semantic-ui-react';
+import { useEffect, useState, Dispatch, SetStateAction } from 'react';
+import { Button, Checkbox, Modal, Form, Radio } from 'semantic-ui-react';
 import CandidateDeciderAPI from '../../API/CandidateDeciderAPI';
 import ResponsesPanel from './ResponsesPanel';
 import LocalProgressPanel from './LocalProgressPanel';
-import GlobalProgressPanel from './GlobalProgressPanel';
 import { useHasAdminPermission, useSelf } from '../Common/FirestoreDataProvider';
 import styles from './CandidateDecider.module.css';
 import SearchBar from './SearchBar';
@@ -16,11 +15,39 @@ type CandidateDeciderProps = {
   uuid: string;
 };
 
+const ratings = [
+  { value: 1, text: 'Strong No', color: 'red' },
+  { value: 2, text: 'No', color: 'orange' },
+  { value: 3, text: 'Maybe', color: 'yellow' },
+  { value: 4, text: 'Yes', color: 'green' },
+  { value: 5, text: 'Strong Yes', color: 'green ' },
+  { value: 0, text: 'Undecided', color: 'grey' }
+];
+
+type CommentEditorProps = {
+  currentComment: string | undefined;
+  setCurrentComment: Dispatch<SetStateAction<string | undefined>>;
+};
+
+const CommentEditor: React.FC<CommentEditorProps> = ({ currentComment, setCurrentComment }) => (
+  <div style={{ width: '100%' }}>
+    <h4>Comments</h4>
+    <Form.Group inline>
+      <Form.Input
+        style={{ height: 256, width: '100%' }}
+        className="fifteen wide field"
+        placeholder={'Comment...'}
+        onChange={(event) => setCurrentComment(event.target.value)}
+        value={currentComment}
+      />
+    </Form.Group>
+  </div>
+);
+
 const CandidateDecider: React.FC<CandidateDeciderProps> = ({ uuid }) => {
   const [isModalOpen, setIsModalOpen] = useState(false);
   const [nextCandidate, setNextCandidate] = useState<number | null>(null);
   const [currentCandidate, setCurrentCandidate] = useState<number>(0);
-  const [showOtherVotes, setShowOtherVotes] = useState<boolean>(false);
   const [seeApplicantName, setSeeApplicantName] = useState<boolean>(false);
 
   const userInfo = useSelf();
@@ -32,8 +59,7 @@ const CandidateDecider: React.FC<CandidateDeciderProps> = ({ uuid }) => {
     const rating = reviews.find(
       (rt) => rt.reviewer.email === userInfo?.email && rt.candidateId === candidate
     );
-    if (rating) return rating.rating;
-    return undefined;
+    return rating ? rating.rating : undefined;
   };
 
   const getComment = (candidate: number) => {
@@ -53,8 +79,8 @@ const CandidateDecider: React.FC<CandidateDeciderProps> = ({ uuid }) => {
     currentComment === defaultCurrentComment && currentRating === defaultCurrentRating;
 
   const populateReviewForCandidate = (candidate: number) => {
-    const rating = getRating(candidate);
-    const comment = getComment(candidate);
+    const rating = getRating(candidate) ?? 0;
+    const comment = getComment(candidate) ?? '';
     setCurrentRating(rating);
     setCurrentComment(comment);
     setDefaultCurrentRating(rating);
@@ -154,87 +180,102 @@ const CandidateDecider: React.FC<CandidateDeciderProps> = ({ uuid }) => {
           </Button>
         </Modal.Actions>
       </Modal>
-      <div className={styles.applicationContainer}>
-        <div className={styles.searchBar}>
-          <SearchBar
-            instance={instance}
-            setCurrentCandidate={(candidate) => {
-              handleCandidateChange(candidate);
-            }}
-            currentCandidate={currentCandidate}
-            seeApplicantName={seeApplicantName}
-          />
+      <div className={styles.leftColumn}>
+        <div className={styles.top}>
+          <div className={styles.navigation}>
+            <div className={styles.progressContainer}>
+              <LocalProgressPanel
+                candidates={instance.candidates}
+                currentCandidate={currentCandidate}
+                reviews={completedReviews}
+              />
+            </div>
+
+            <div className={styles.searchBar}>
+              <Button
+                basic
+                color="blue"
+                disabled={currentCandidate === 0}
+                onClick={() => {
+                  handleCandidateChange(currentCandidate - 1);
+                }}
+              >
+                PREVIOUS
+              </Button>
+              <SearchBar
+                instance={instance}
+                setCurrentCandidate={(candidate) => {
+                  handleCandidateChange(candidate);
+                }}
+                currentCandidate={currentCandidate}
+                seeApplicantName={seeApplicantName}
+              />
+              <Button.Group className={styles.previousNextButtonContainer}>
+                <Button
+                  basic
+                  color="blue"
+                  disabled={currentCandidate === instance.candidates.length - 1}
+                  onClick={() => {
+                    handleCandidateChange(currentCandidate + 1);
+                  }}
+                >
+                  NEXT
+                </Button>
+              </Button.Group>
+            </div>
+          </div>
+          <div className={styles.commentEditorWrapper}>
+            <CommentEditor currentComment={currentComment} setCurrentComment={setCurrentComment} />
+          </div>
         </div>
-        <div className={styles.controlsContainer}>
-          <h4 className={styles.candidateIDTitle}>Candidate ID:</h4>
-          <Dropdown
-            compact
-            value={currentCandidate}
-            selection
-            options={instance.candidates.map((candidate) => ({
-              value: candidate.id,
-              key: candidate.id,
-              text: candidate.id + 1 // offset by 1 to account for 0-indexed array
-            }))}
-            onChange={(_, data) => {
-              handleCandidateChange(data.value as number);
-            }}
-          />
-          <span className={styles.ofNum}>of {instance.candidates.length}</span>
-          <Button.Group className={styles.previousNextButtonContainer}>
+        <div className={styles.bottom}>
+          <div className={styles.ratingSelectorWrapper}>
+            <h4>Final selection</h4>
+
+            <Form>
+              <Form.Group inline>
+                {ratings.map((rt) => (
+                  <Form.Field key={rt.value}>
+                    <Radio
+                      label={rt.text}
+                      name="rating-group"
+                      value={rt.value}
+                      color={rt.color}
+                      checked={rt.value === currentRating}
+                      onClick={() => setCurrentRating(rt.value as Rating)}
+                    />
+                  </Form.Field>
+                ))}
+              </Form.Group>
+            </Form>
+          </div>
+          <div className={styles.saveButtonWrapper}>
             <Button
-              basic
-              color="blue"
-              disabled={currentCandidate === 0}
+              className="ui blue button"
+              disabled={isSaved}
               onClick={() => {
-                handleCandidateChange(currentCandidate - 1);
+                handleRatingAndCommentChange(
+                  currentCandidate,
+                  currentRating ?? 0,
+                  currentComment ?? ''
+                );
               }}
             >
-              PREVIOUS
+              Save
             </Button>
-            <Button
-              basic
-              color="blue"
-              disabled={currentCandidate === instance.candidates.length - 1}
-              onClick={() => {
-                handleCandidateChange(currentCandidate + 1);
-              }}
-            >
-              NEXT
-            </Button>
-          </Button.Group>
-          <Button
-            className="ui blue button"
-            disabled={isSaved}
-            onClick={() => {
-              handleRatingAndCommentChange(
-                currentCandidate,
-                currentRating ?? 0,
-                currentComment ?? ''
-              );
-            }}
-          >
-            Save
-          </Button>
-          {hasAdminPermission && (
-            <Checkbox
-              className={styles.showOtherVotes}
-              toggle
-              checked={showOtherVotes}
-              onChange={() => setShowOtherVotes((prev) => !prev)}
-              label="Show other people's votes"
-            />
-          )}
-          {hasAdminPermission && (
-            <Checkbox
-              className={styles.seeApplicantName}
-              toggle
-              checked={seeApplicantName}
-              onChange={() => setSeeApplicantName((prev) => !prev)}
-              label="See applicant name"
-            />
-          )}
+          </div>
         </div>
+      </div>
+      <div className={styles.responsesContainer}>
+        {hasAdminPermission && (
+          <Checkbox
+            className={styles.seeApplicantName}
+            toggle
+            checked={seeApplicantName}
+            onChange={() => setSeeApplicantName((prev) => !prev)}
+            label="See applicant name"
+          />
+        )}
         <ResponsesPanel
           headers={instance.headers}
           responses={instance.candidates[currentCandidate].responses}
@@ -244,19 +285,6 @@ const CandidateDecider: React.FC<CandidateDeciderProps> = ({ uuid }) => {
           setCurrentRating={setCurrentRating}
           seeApplicantName={seeApplicantName}
           candidate={instance.candidates[currentCandidate].id}
-        />
-      </div>
-      <div className={styles.progressContainer}>
-        <LocalProgressPanel
-          showOtherVotes={showOtherVotes}
-          candidates={instance.candidates}
-          currentCandidate={currentCandidate}
-          reviews={completedReviews}
-        />
-        <GlobalProgressPanel
-          showOtherVotes={showOtherVotes}
-          candidates={instance.candidates}
-          reviews={completedReviews}
         />
       </div>
     </div>
