+++ frontend/src/components/Candidate-Decider/GlobalProgressPanel.tsx
@@ -22,7 +22,6 @@ const GlobalProgressPanel: React.FC<Props> = ({ showOtherVotes, candidates, revi
         size="tiny"
         color="blue"
       >{`${reviews.length}/${candidates.length * allReviewers.length}`}</Progress>
-      <RatingsDisplay ratings={reviews} header="Global Rating Statistics" />
       {showOtherVotes ? (
         <>
           <h3>Per-person Ratings</h3>
