+++ frontend/src/components/Candidate-Decider/LocalProgressPanel.tsx
@@ -6,14 +6,12 @@ import RatingsDisplay from './RatingsDisplay';
 import { ratingToString } from './ratings-utils';
 
 type ProgressPanelProps = {
-  showOtherVotes: boolean;
   candidates: CandidateDeciderCandidate[];
   currentCandidate: number;
   reviews: CandidateDeciderReview[];
 };
 
 const LocalProgressPanel: React.FC<ProgressPanelProps> = ({
-  showOtherVotes,
   candidates,
   currentCandidate,
   reviews
@@ -33,8 +31,8 @@ const LocalProgressPanel: React.FC<ProgressPanelProps> = ({
         size="tiny"
         color="blue"
       >{`${myRatings.length}/${candidates.length}`}</Progress>
-      <RatingsDisplay ratings={myRatings} header="My Rating Statistics" />
-      {showOtherVotes && userInfo && LEAD_ROLES.includes(userInfo.role) ? (
+
+      {userInfo && LEAD_ROLES.includes(userInfo.role) ? (
         <>
           <RatingsDisplay
             ratings={currentCandidateReviews}
