+++ frontend/src/components/Candidate-Decider/ProgressPanel.module.css
@@ -1,6 +1,5 @@
 .progressContainer {
   width: 100%;
-  padding: 2%;
 }
 
 .percentageBar {
